fix(build): fix docs.rs build failure
The docs.rs build was failing with error E0753 because inner doc comments
(//\!) were used in the middle of generated bindings file. Inner doc comments
can only appear at the beginning of a file or module.

Changed to regular comments (//) which won't be exposed in documentation
and won't cause compilation errors.